etl_rds_to_s3/requirements.txt
pylint
pytest
pyodbc
python-dotenv
pandas